Add ShelterDAO object to access database. Implement getNearest function in LocationService. Change button click in main page jsp. Still cannot be used because it hasn't handle shelter list return value of the JSON. Add 2 beans in servlet-context for LocationService and ShelterDAO.

+++ src/main/java/com/roxinlabs/transjakarta/services/LocationService.java
@@ -9,5 +9,5 @@ public interface LocationService {
 	
 	public Foursquare getLocationList(String query);
 	
-	public List<Shelter> getNearest(int lat, int lng);
+	public List<Shelter> getNearest(Double lat, Double lng);
 }
